pyproject.toml
[project]
name = "facerec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "rawpy>=0.23.2",
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.36",
    "scikit-image>=0.24.0",
    "exifread>=3.0.0",
    "lancedb>=0.10.2",
    "numpy<2.0.0",
    "alembic>=1.13.3",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.20.0",
    "torch>=2.5.1",
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "python-multipart>=0.0.19",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "fuzzywuzzy[speedup]>=0.18.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "icecream>=2.1.3",
    "streamlit>=1.39.0",
]
